back-end/requirements.txt
#ASGI Server
uvicorn
#API framework
flask
flask_cors
#Image manipulation
pillow
#Artificial Intelligence
